+++ src/app/my/_hooks/queries/useReviewLabelList.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { reviewLabelList } from "../../api/mypage";
+import { reviewLabelList } from "../../../api/mypage";
 
 const useReviewLabelList = (userId: number) => {
   return useQuery({
