+++ src/app/my/_hooks/queries/useReviewList.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { reviewList } from "../../api/mypage";
+import { reviewList } from "../../../api/mypage";
 
 const useReviewList = (userId: number) => {
   return useQuery({
